refactor: colocate two-factor service files and their tests
