+++ app/lib/methods/helpers/handleSaveUserProfileError.ts
@@ -1,6 +1,6 @@
 import I18n from '../../../i18n';
 import { showErrorAlert } from '.';
-import { isTwoFactorCancelled } from '../../services/twoFactorCancelled';
+import { isTwoFactorCancelled } from '../../services/twoFactor/twoFactorCancelled';
 
 const handleSaveUserProfileError = (e: any, action: string) => {
 	if (isTwoFactorCancelled(e)) {
