+++ app/views/E2EEToggleRoomView/resetRoomKey.ts
@@ -5,7 +5,7 @@ import { Encryption } from '../../lib/encryption';
 import log from '../../lib/methods/helpers/log';
 import { showToast } from '../../lib/methods/helpers/showToast';
 import { e2eResetRoomKey } from '../../lib/services/restApi';
-import { isTwoFactorCancelled } from '../../lib/services/twoFactorCancelled';
+import { isTwoFactorCancelled } from '../../lib/services/twoFactor/twoFactorCancelled';
 
 export const resetRoomKey = (rid: string) => {
 	Alert.alert(
